requirements.txt
Flask>=3.0,<4
Flask-SQLAlchemy>=3.1,<4
SQLAlchemy>=2.0,<3
PyMySQL>=1.1,<2
Flask-Login>=0.6,<1
python-dotenv>=1.0,<2
